# lib/api-zod/src/generated/types/groupStanding.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * FIFA Hub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StandingRow } from './standingRow';

export interface GroupStanding {
  group: string;
  standings: StandingRow[];
}