# lib/api-zod/src/generated/types/standingRow.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * FIFA Hub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Team } from './team';

export interface StandingRow {
  rank: number;
  team: Team;
  played: number;
  wins?: number;
  draws?: number;
  losses?: number;
  goalsFor?: number;
  goalsAgainst?: number;
  goalDifference?: number;
  points: number;
}