# lib/api-zod/src/generated/types/teamDetail.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * FIFA Hub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Fixture } from './fixture';
import type { NewsArticle } from './newsArticle';
import type { Player } from './player';
import type { TeamDetailHistoricalStats } from './teamDetailHistoricalStats';
import type { TeamDetailHistoryTimelineItem } from './teamDetailHistoryTimelineItem';
import type { TeamInsight } from './teamInsight';
import type { TeamLineup } from './teamLineup';
import type { TeamStats } from './teamStats';

export interface TeamDetail {
  id: number;
  name: string;
  logo: string;
  /** @nullable */
  code?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  group: string | null;
  /** @nullable */
  fifaRank: number | null;
  /** @nullable */
  coach?: string | null;
  /** @nullable */
  founded?: number | null;
  /** @nullable */
  venue?: string | null;
  players?: Player[];
  fixtures?: Fixture[];
  statistics?: TeamStats;
  /** @nullable */
  manager?: string | null;
  /** @nullable */
  captain?: string | null;
  /** @nullable */
  qualifiedVia?: string | null;
  /** @nullable */
  appearances?: number | null;
  /** @nullable */
  confederation?: string | null;
  /** @nullable */
  historicalStats?: TeamDetailHistoricalStats;
  historyTimeline?: TeamDetailHistoryTimelineItem[];
  insights?: TeamInsight[];
  news?: NewsArticle[];
  lineup?: TeamLineup;
}